fastApi/requirements.txt
fastapi
uvicorn
pandas
python-jobspy
google-generativeai
pydantic
python-multipart
boto3